+++ QuickFIXn/Session.cs
@@ -80,11 +80,11 @@ namespace QuickFix
         {
             get
             {
-                return state_.GetNextSenderMsgSeqNum();
+                return state_.NextSenderMsgSeqNum;
             }
             set
             {
-                state_.SetNextSenderMsgSeqNum(value);
+                state_.NextSenderMsgSeqNum = value;
             }
         }
 
@@ -95,11 +95,11 @@ namespace QuickFix
         {
             get
             {
-                return state_.GetNextTargetMsgSeqNum();
+                return state_.NextTargetMsgSeqNum;
             }
             set
             {
-                state_.SetNextTargetMsgSeqNum(value);
+                state_.NextTargetMsgSeqNum = value;
             }
         }
 
@@ -786,13 +786,13 @@ namespace QuickFix
 
                     if ((endSeqNo == 999999) || (endSeqNo == 0))
                     {
-                        endSeqNo = state_.GetNextSenderMsgSeqNum() - 1;
+                        endSeqNo = state_.NextSenderMsgSeqNum - 1;
                     }
 
                     if (!PersistMessages)
                     {
                         endSeqNo++;
-                        int next = state_.GetNextSenderMsgSeqNum();
+                        int next = state_.NextSenderMsgSeqNum;
                         if (endSeqNo > next)
                             endSeqNo = next;
                         GenerateSequenceReset(resendReq, begSeqNo, endSeqNo);
@@ -845,7 +845,7 @@ namespace QuickFix
                         current = msgSeqNum + 1;
                     }
 
-                    int nextSeqNum = state_.GetNextSenderMsgSeqNum();
+                    int nextSeqNum = state_.NextSenderMsgSeqNum;
                     if (++endSeqNo > nextSeqNum)
                     {
                         endSeqNo = nextSeqNum;
@@ -932,15 +932,15 @@ namespace QuickFix
             if (sequenceReset.IsSetField(Fields.Tags.NewSeqNo))
             {
                 int newSeqNo = sequenceReset.GetInt(Fields.Tags.NewSeqNo);
-                this.Log.OnEvent("Received SequenceReset FROM: " + state_.GetNextTargetMsgSeqNum() + " TO: " + newSeqNo);
+                this.Log.OnEvent("Received SequenceReset FROM: " + state_.NextTargetMsgSeqNum + " TO: " + newSeqNo);
 
-                if (newSeqNo > state_.GetNextTargetMsgSeqNum())
+                if (newSeqNo > state_.NextTargetMsgSeqNum)
                 {
-                    state_.SetNextTargetMsgSeqNum(newSeqNo);
+                    state_.NextTargetMsgSeqNum = newSeqNo;
                 }
                 else
                 {
-                    if (newSeqNo < state_.GetNextTargetMsgSeqNum())
+                    if (newSeqNo < state_.NextTargetMsgSeqNum)
                         GenerateReject(sequenceReset, FixValues.SessionRejectReason.VALUE_IS_INCORRECT);
                 }
             }
@@ -1077,8 +1077,8 @@ namespace QuickFix
         {
             return (this.SessionID.BeginString.CompareTo(FixValues.BeginString.FIX41) >= 0)
                 && (this.ResetOnLogon || this.ResetOnLogout || this.ResetOnDisconnect)
-                && (state_.GetNextSenderMsgSeqNum() == 1)
-                && (state_.GetNextTargetMsgSeqNum() == 1);
+                && (state_.NextSenderMsgSeqNum == 1)
+                && (state_.NextTargetMsgSeqNum == 1);
         }
 
         protected bool IsCorrectCompID(string senderCompID, string targetCompID)
@@ -1097,19 +1097,19 @@ namespace QuickFix
 
         protected bool IsTargetTooHigh(int msgSeqNum)
         {
-            return msgSeqNum > state_.GetNextTargetMsgSeqNum();
+            return msgSeqNum > state_.NextTargetMsgSeqNum;
         }
 
         protected bool IsTargetTooLow(int msgSeqNum)
         {
-            return msgSeqNum < state_.GetNextTargetMsgSeqNum();
+            return msgSeqNum < state_.NextTargetMsgSeqNum;
         }
 
         protected void DoTargetTooHigh(Message msg, int msgSeqNum)
         {
             string beginString = msg.Header.GetString(Fields.Tags.BeginString);
 
-            this.Log.OnEvent("MsgSeqNum too high, expecting " + state_.GetNextTargetMsgSeqNum() + " but received " + msgSeqNum);
+            this.Log.OnEvent("MsgSeqNum too high, expecting " + state_.NextTargetMsgSeqNum + " but received " + msgSeqNum);
             state_.Queue(msgSeqNum, msg);
 
             if (state_.ResendRequested())
@@ -1134,7 +1134,7 @@ namespace QuickFix
 
             if (!possDupFlag)
             {
-                string err = "MsgSeqNum too low, expecting " + state_.GetNextTargetMsgSeqNum() + " but received " + msgSeqNum;
+                string err = "MsgSeqNum too low, expecting " + state_.NextTargetMsgSeqNum + " but received " + msgSeqNum;
                 GenerateLogout(err);
                 throw new QuickFIXException(err);
             }
@@ -1224,7 +1224,7 @@ namespace QuickFix
 
         protected bool GenerateResendRequest(string beginString, int msgSeqNum)
         {
-            int beginSeqNum = state_.GetNextTargetMsgSeqNum();
+            int beginSeqNum = state_.NextTargetMsgSeqNum;
             int endRangeSeqNum = msgSeqNum - 1;
             int endChunkSeqNum;
             if (this.MaxMessagesInResendRequest > 0)
@@ -1436,7 +1436,7 @@ namespace QuickFix
             }
             if (!MsgType.LOGON.Equals(msgType)
               && !MsgType.SEQUENCE_RESET.Equals(msgType)
-              && (msgSeqNum == state_.GetNextTargetMsgSeqNum()))
+              && (msgSeqNum == state_.NextTargetMsgSeqNum))
             {
                 state_.IncrNextTargetMsgSeqNum();
             }
@@ -1511,7 +1511,7 @@ namespace QuickFix
             if (msgSeqNum > 0)
                 m.Header.SetField(new Fields.MsgSeqNum(msgSeqNum));
             else
-                m.Header.SetField(new Fields.MsgSeqNum(state_.GetNextSenderMsgSeqNum()));
+                m.Header.SetField(new Fields.MsgSeqNum(state_.NextSenderMsgSeqNum));
 
             if (this.EnableLastMsgSeqNumProcessed && !m.Header.IsSetField(Tags.LastMsgSeqNumProcessed))
             {
@@ -1653,7 +1653,7 @@ namespace QuickFix
                         if (resetSeqNumFlag.getValue())
                         {
                             state_.Reset("ResetSeqNumFlag");
-                            message.Header.SetField(new Fields.MsgSeqNum(state_.GetNextSenderMsgSeqNum()));
+                            message.Header.SetField(new Fields.MsgSeqNum(state_.NextSenderMsgSeqNum));
                         }
                         state_.SentReset = resetSeqNumFlag.Obj;
                     }