+++ RELEASE_NOTES.md
@@ -17,7 +17,8 @@ it technically violates semantic versioning.
 * #748 - Update all csproj files to net6.0, fix .NET deprecations, update pwsh scripts, remove some non-pwsh scripts (gbirchmeier)
 * #749 - All deprecations are removed (gbirchmeier)
 * #746 - remove some unused public `FIXnn_LAST_FIELD` constants; corrected FIX44 DD `UNKOWN_ID` typo (gbirchmeier)
-* TBD - privatize DDField attribute setters (gbirchmeier)
+* #756 - privatize DDField property setters,
+         change various Get/SetNextSenderMsgSeqNum & Get/SetNextTargetMsgSeqNum functions to properties (gbirchmeier)
 
 **Non-breaking changes**
 * (patch) #647 - replace lock with memory barrier to avoid deadlocks (brunobelmondo)